requirements.txt
pandas>=2.2.2
telethon==1.28.5
psycopg2-binary==2.9.9
textblob==0.17.1
sqlalchemy==2.0.20
python-dotenv==1.0.0